Redesign landing page
This combines the splash screen and wallet setup page into one.
The wallet setup buttons show up if there's no wallet and the normal
loading process continues if there's at least one wallet.
