+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
         tools:ignore="GoogleAppIndexingWarning" >
 
         <activity
-            android:name=".activities.SplashScreen"
+            android:name=".activities.SplashScreenActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:label="@string/app_name"
             android:launchMode="singleTask">
