+++ app/src/main/res/drawable/transparent_ripple_corner_8dp.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2018-2019 The Decred developers
+  ~ Use of this source code is governed by an ISC
+  ~ license that can be found in the LICENSE file.
+  -->
+
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_enabled="false">
+        <shape android:shape="rectangle">
+            <solid android:color="@color/lightGray"/>
+            <corners android:radius="8dp"/>
+        </shape>
+    </item>
+
+    <item android:state_enabled="true">
+        <ripple android:color="@color/primaryRipple">
+            <item android:id="@android:id/mask">
+                <shape android:shape="rectangle">
+                    <solid android:color="@android:color/white" />
+                    <corners
+                        android:radius="8dp"/>
+                </shape>
+            </item>
+
+            <item>
+                <shape android:shape="rectangle">
+                    <solid android:color="@android:color/transparent"/>
+                    <corners android:radius="8dp"/>
+                </shape>
+            </item>
+
+        </ripple>
+    </item>
+
+</selector>
