+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -1,0 +1,142 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2018-2019 The Decred developers
+  ~ Use of this source code is governed by an ISC
+  ~ license that can be found in the LICENSE file.
+  -->
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:background="@color/colorPrimary"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        android:paddingStart="@dimen/margin_padding_size_24"
+        android:paddingEnd="@dimen/margin_padding_size_24"
+        android:clipToPadding="true"
+        android:gravity="center"
+        android:orientation="vertical">
+
+        <ImageView
+            android:id="@+id/splashscreen_dcr_symbol"
+            android:layout_width="160dp"
+            android:layout_height="160dp"
+            android:src="@drawable/avd_anim" />
+
+        <TextView
+            android:id="@+id/tv_testnet"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/testnet"
+            android:textColor="@color/darkBlueTextColor"
+            android:fontFamily="@font/source_sans_pro"
+            android:textSize="20sp"
+            android:textStyle="bold"
+            android:visibility="gone"
+            tools:visibility="visible" />
+
+        <TextView
+            android:id="@+id/loading_status"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="16dp"
+            android:text="@string/loading"
+            android:textColor="@color/darkBlueTextColor"
+            android:textSize="19sp"
+            app:fontFamily="@font/source_sans_pro_regular" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/welcome_text"
+            android:layout_marginTop="@dimen/margin_padding_size_24"
+            android:textColor="@color/darkBlueTextColor"
+            android:textSize="@dimen/edit_text_size_24"
+            android:text="@string/welcome_screen_text"
+            android:fontFamily="@font/source_sans_pro"
+            android:visibility="gone"
+            android:gravity="center" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/bottom_bar_layout"
+        android:visibility="gone"
+        android:padding="@dimen/margin_padding_size_12"
+        android:orientation="vertical">
+
+        <LinearLayout
+            android:id="@+id/ll_create_wallet"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/margin_padding_size_8"
+            android:background="@drawable/blue_ripple_corners_8dp"
+            android:clickable="true"
+            android:elevation="@dimen/shadow_spread"
+            android:focusable="true"
+            android:gravity="center_vertical"
+            android:orientation="horizontal"
+            android:padding="@dimen/margin_padding_size_16">
+
+            <ImageView
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                app:srcCompat="@drawable/ic_create_wallet_24px" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="@dimen/margin_padding_size_16"
+                android:includeFontPadding="false"
+                android:text="@string/create_a_new_wallet"
+                android:textColor="@color/white"
+                android:textSize="@dimen/edit_text_size_16"
+                app:fontFamily="@font/source_sans_pro" />
+
+        </LinearLayout>
+
+        <LinearLayout
+            android:id="@+id/ll_restore_wallet"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="@drawable/transparent_ripple_corner_8dp"
+            android:padding="@dimen/margin_padding_size_16"
+            android:layout_marginTop="@dimen/margin_padding_size_16"
+            android:layout_marginBottom="@dimen/margin_padding_size_8"
+            android:layout_marginLeft="@dimen/margin_padding_size_8"
+            android:layout_marginRight="@dimen/margin_padding_size_8"
+            android:gravity="center_vertical"
+            android:elevation="@dimen/shadow_spread"
+            android:clickable="true"
+            android:focusable="true"
+            android:orientation="horizontal">
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_vertical"
+                app:srcCompat="@drawable/ic_restore_wallet_24px" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/restore_existing_wallet"
+                android:layout_marginStart="@dimen/margin_padding_size_16"
+                android:textColor="@color/blue"
+                android:textSize="@dimen/edit_text_size_18"
+                android:includeFontPadding="false"
+                app:fontFamily="@font/source_sans_pro" />
+
+        </LinearLayout>
+
+    </LinearLayout>
+
+</LinearLayout>
